Rewrite all low-level API bindings using a helper macro.
This gives two very nice properties:

- Now changes to the underlying FFI are contained to the macro.
- The resulting bindings can actually be used on their own for
  low-level FFI buffer usage, because all the ugly temporary
  object allocation stuff is hidden. Some of them completely
  replace the old wrapper functions.

Also, the test-* functions now don't return the flag argument,
because its value can be effectively communicated by the fact
that other return values are nil or non-nil. In turn, the other
values cannot safely be non-nil when flag is false because
in that case their values are garbage by definition.