Implement status callbacks for requests.
When a wait or test API function wrapper obtains a valid
status for a request, it calls this callback. It can be
used to implement automatic cleanup of temporary objects
allocated to support an asynchronous operation.

Also, the status is now set to NIL if an MPI_UNDEFINED
index is returned by the underlying function.