Stop memoizing the result of get-typespec-by-subtype.
It is most often applied to the result of type-of on
various values, which has a tendency to include the
value in the type, e.g. (integer 3 3). Thus, caching
would in fact cause a memory leak.

In order to reduce the inefficiency, avoid searching
for subtype before trying stringp and arrayp.